backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
urllib3==2.5.0
